lesson 24

+++ app/Http/Controllers/ThreadsController.php
@@ -109,6 +109,13 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        if ($thread->user_id != auth()->id()){
+            abort(403, 'You do not have permission to do this');
+        }
+
+
         // lesson 23 cascade delete 
         // $thread->replies()->delete();
         $thread->delete();
